Move dump logic to separate file
