+++ crates/starknet/src/starknet/mod.rs
@@ -1,9 +1,6 @@
 use std::collections::HashMap;
 use std::fmt;
-use std::fs::{self, File};
-use std::io::Read;
 use std::net::IpAddr;
-use std::path::Path;
 use std::time::SystemTime;
 
 use starknet_api::block::{BlockNumber, BlockStatus, BlockTimestamp, GasPrice};
@@ -40,8 +37,7 @@ use starknet_types::rpc::transactions::broadcasted_declare_transaction_v2::Broad
 use starknet_types::rpc::transactions::broadcasted_deploy_account_transaction::BroadcastedDeployAccountTransaction;
 use starknet_types::rpc::transactions::broadcasted_invoke_transaction::BroadcastedInvokeTransaction;
 use starknet_types::rpc::transactions::{
-    BroadcastedTransaction, BroadcastedTransactionCommon, DeclareTransaction, InvokeTransaction,
-    Transaction, Transactions,
+    BroadcastedTransaction, BroadcastedTransactionCommon, Transaction, Transactions,
 };
 use starknet_types::traits::HashProducer;
 use strum_macros::EnumIter;
@@ -69,6 +65,7 @@ use crate::transactions::{StarknetTransaction, StarknetTransactions};
 mod add_declare_transaction;
 mod add_deploy_account_transaction;
 mod add_invoke_transaction;
+mod dump;
 mod estimations;
 mod events;
 mod get_class_impls;
@@ -194,78 +191,6 @@ impl Starknet {
         }
 
         Ok(this)
-    }
-
-    pub fn re_execute(&mut self, transactions: StarknetTransactions) -> DevnetResult<()> {
-        for (_, transaction) in transactions.iter() {
-            match transaction.inner.clone() {
-                Transaction::Declare(DeclareTransaction::Version0(_)) => {
-                    return Err(Error::SerializationNotSupported);
-                }
-                Transaction::Declare(DeclareTransaction::Version1(tx)) => {
-                    let contract_class = self
-                        .state
-                        .contract_classes
-                        .get(&tx.class_hash)
-                        .ok_or(Error::ContractClassLoadError)?;
-                    if let ContractClass::Cairo0(contract) = contract_class {
-                        let declare_tx = BroadcastedDeclareTransactionV1::new(
-                            tx.sender_address,
-                            tx.max_fee,
-                            &tx.signature,
-                            tx.nonce,
-                            contract,
-                            tx.version,
-                        );
-                        self.add_declare_transaction_v1(declare_tx)?;
-                    } else {
-                        return Err(Error::SerializationNotSupported);
-                    };
-                }
-                Transaction::Declare(DeclareTransaction::Version2(tx)) => {
-                    let declare_tx = BroadcastedDeclareTransactionV2::new(
-                        &tx.contract_class,
-                        tx.compiled_class_hash,
-                        tx.sender_address,
-                        tx.max_fee,
-                        &tx.signature,
-                        tx.nonce,
-                        tx.version,
-                    );
-                    self.add_declare_transaction_v2(declare_tx)?;
-                }
-                Transaction::DeployAccount(tx) => {
-                    let deploy_account_tx = BroadcastedDeployAccountTransaction::new(
-                        &tx.constructor_calldata,
-                        tx.max_fee,
-                        &tx.signature,
-                        tx.nonce,
-                        tx.class_hash,
-                        tx.contract_address_salt,
-                        tx.version,
-                    );
-                    self.add_deploy_account_transaction(deploy_account_tx)?;
-                }
-                Transaction::Deploy(_) => return Err(Error::SerializationNotSupported),
-                Transaction::Invoke(InvokeTransaction::Version0(_)) => {
-                    return Err(Error::SerializationNotSupported);
-                }
-                Transaction::Invoke(InvokeTransaction::Version1(tx)) => {
-                    let invoke_tx = BroadcastedInvokeTransaction::new(
-                        tx.sender_address,
-                        tx.max_fee,
-                        &tx.signature,
-                        tx.nonce,
-                        &tx.calldata,
-                        tx.version,
-                    );
-                    self.add_invoke_transaction(invoke_tx)?;
-                }
-                Transaction::L1Handler(_) => return Err(Error::SerializationNotSupported),
-            };
-        }
-        println!("re_execute");
-        Ok(())
     }
 
     pub fn get_predeployed_accounts(&self) -> Vec<Account> {
@@ -524,56 +449,6 @@ impl Starknet {
         invoke_transaction: BroadcastedInvokeTransaction,
     ) -> DevnetResult<TransactionHash> {
         add_invoke_transaction::add_invoke_transaction(self, invoke_transaction)
-    }
-
-    /// save starknet transactions to file
-    pub fn dump_transactions(&self) -> DevnetResult<()> {
-        match &self.config.dump_path {
-            Some(path) => {
-                let starknet_dump = serde_json::to_string(&self.transactions).map_err(|_| {
-                    Error::SerializationError { obj_name: "StarknetTransactions".to_string() }
-                })?;
-                let encoded: Vec<u8> = bincode::serialize(&starknet_dump).map_err(|_| {
-                    Error::SerializationError { obj_name: "StarknetTransactions".to_string() }
-                })?;
-                fs::write(Path::new(&path), encoded)?;
-
-                Ok(())
-            }
-            None => Err(Error::FormatError),
-        }
-    }
-
-    // load starknet transactions from file
-    pub fn load_transactions(&self) -> DevnetResult<StarknetTransactions> {
-        match &self.config.dump_path {
-            Some(path) => {
-                let file_path = Path::new(path);
-
-                // load only if the file exists, if dump_path is set but the file doesn't exist it
-                // can mean that it's first run with dump_path parameter set to dump, in that case
-                // return default value of StarknetTransactions
-                if file_path.exists() {
-                    let mut file = File::open(file_path)?;
-                    let mut v: Vec<u8> = Vec::new();
-                    file.read_to_end(&mut v)?;
-                    let decoded: Result<String, Error> = bincode::deserialize(&v).map_err(|_| {
-                        Error::DeserializationError { obj_name: "StarknetTransactions".to_string() }
-                    });
-                    let transactions: DevnetResult<StarknetTransactions, Error> =
-                        serde_json::from_str(decoded.unwrap().as_str()).map_err(|_| {
-                            Error::DeserializationError {
-                                obj_name: "StarknetTransactions".to_string(),
-                            }
-                        });
-
-                    transactions
-                } else {
-                    Ok(StarknetTransactions::default())
-                }
-            }
-            None => Err(Error::FormatError),
-        }
     }
 
     /// Creates an invoke tx for minting, using the chargeable account.
